requirements.txt
textual>=0.47.1
telethon>=1.34.0
cryptg>=0.4.0
python-dotenv>=1.0.0
aiofiles>=23.2.1
